examples/investment_analysis/requirements.txt
crewai
composio-crewai
composio-langchain
langchain-openai
python-dotenv